Create deconvolution module, and refactor it somewhat. Fix deskew_yaml test
